client DAO pas fini

+++ src/main/java/com/epf/rentmanager/model/Client.java
@@ -1,13 +1,13 @@
 package com.epf.rentmanager.model;
 import java.time.LocalDate;
 public class Client {
-    private int id;
+    private long id;
     private String nom;
     private String prenom;
     private String email;
     private LocalDate naissance;
 
-    public Client(int id, String nom, String prenom, String email, LocalDate naissance) {
+    public Client(long id, String nom, String prenom, String email, LocalDate naissance) {
         this.id = id;
         this.nom = nom;
         this.prenom = prenom;
@@ -15,7 +15,7 @@ public class Client {
         this.naissance = naissance;
     }
 
-    public int getId() {
+    public long getId() {
         return this.id;
     }
